ENH: The DCDFile write method now requires the double type for box and unit tests have been adjusted accordingly, based on reviewer comments in PR#1155.

+++ package/MDAnalysis/lib/formats/libdcd.pyx
@@ -346,7 +346,7 @@ cdef class DCDFile:
         if ok != 0:
             raise IOError("Writing DCD header failed: {}".format(DCD_ERRORS[ok]))
 
-    def write(self, xyz, float [:] box, int step, float time, int natoms, int charmm):
+    def write(self, xyz, double [:] box, int step, float time, int natoms, int charmm):
         """write one frame into DCD file.
 
         Parameters
